@@ -11,30 +11,24 @@
         "type": "git",
         "url": "https://github.com/girder/large_image.git"
     },
-    "peerDependencies": {
-        "@girder/core": "*"
-    },
+    "type": "module",
     "dependencies": {
-        "codemirror": "^5.65.0",
-        "copy-webpack-plugin": "^4.5.2",
-        "d3": "^3.5.16",
-        "geojs": "^1.8.6",
         "@egjs/hammerjs": "^2.0.8",
+        "@girder/core": "^5.0.0-a10",
+        "codemirror": "^5.65.0",
+        "d3": "^7.9.0",
+        "geojs": "^1.8.6",
         "js-yaml": "^4.1.0",
         "jsonlint-mod": "^1.7.6",
+        "nib": "^1.2.0",
         "sinon": "^2.1.0",
-        "vue": "~2.6.14",
+        "vue": "^2.7.16",
+        "vue-color": "^2.8.1",
         "vue-loader": "~15.9.8",
         "vue-template-compiler": "~2.6.14",
-        "webpack": "^3",
         "yaml": "^2.1.1"
     },
     "main": "./index.js",
-    "girderPlugin": {
-        "name": "large_image",
-        "main": "./main.js",
-        "webpack": "webpack.helper"
-    },
     "eslintConfig": {
         "extends": [
             "@girder",
@@ -69,7 +63,13 @@
             "vue/multiline-html-element-content-newline": "off",
             "vue/html-self-closing": "off"
         },
-        "root": true
+        "ignorePatterns": [
+            "dist/**"
+        ],
+        "root": true,
+        "globals": {
+            "girder": "readonly"
+        }
     },
     "eslintIgnore": [
         "**/node_modules/"
@@ -90,6 +90,8 @@
     "devDependencies": {
         "@girder/eslint-config": "*",
         "@girder/pug-lint-config": "*",
+        "@playwright/test": "^1.51.0",
+        "@vitejs/plugin-vue2": "^2.3.1",
         "eslint": "^8.20.0",
         "eslint-config-semistandard": "^17.0.0",
         "eslint-config-standard": "^17.0.0",
@@ -98,11 +100,18 @@
         "eslint-plugin-n": "^15.2.4",
         "eslint-plugin-promise": "^6.0.0",
         "eslint-plugin-vue": "~9.14.0",
+        "playwright": "^1.51.0",
+        "pug": "^3.0.3",
         "pug-lint": "^2.6.0",
-        "stylus-supremacy": "^2.17.5"
+        "stylus-supremacy": "^2.17.5",
+        "vite": "^5.2.13",
+        "vite-plugin-static-copy": "^1.0.6"
     },
     "scripts": {
+        "build": "vite build",
+        "watch": "vite build --watch",
         "lint": "eslint --cache . && eslint --no-eslintrc -c ./package.json --cache ../../test_girder/web_client_specs && pug-lint . && stylus-supremacy format --compare ./**/*.styl --options package.json",
-        "format": "eslint --cache --fix . && eslint --no-eslintrc -c ./package.json --cache --fix ../../test_girder/web_client_specs && stylus-supremacy format ./**/*.styl --replace --options package.json"
+        "format": "eslint --cache --fix . && eslint --no-eslintrc -c ./package.json --cache --fix ../../test_girder/web_client_specs && stylus-supremacy format ./**/*.styl --replace --options package.json",
+        "test": "playwright test"
     }
 }
