@@ -1,5 +1,6 @@
 <script>
-module.exports = {
+// TODO: consider removing all mutations of the prop `currentFrame`
+export default {
     name: 'FrameSelector',
     props: [
         'currentFrame',
@@ -80,12 +81,14 @@ module.exports = {
     },
     methods: {
         updateCurrentFrame(value) {
+            // eslint-disable-next-line
             this.currentFrame = +value;
         },
         setCurrentMode(mode) {
             this.currentModeId = mode.id;
         },
         setCurrentFrame(frame) {
+            // eslint-disable-next-line
             this.currentFrame = +frame;
             this.indexInfo = Object.fromEntries(
                 Object.entries(this.indexInfo)
@@ -129,6 +132,7 @@ module.exports = {
             this.updateHistogramParamStyles();
         },
         updateFrameSlider(frame) {
+            // eslint-disable-next-line
             this.currentFrame = +frame;
             this.frameUpdate({frame, undefined});
         },
@@ -142,6 +146,7 @@ module.exports = {
                     }
                 }
             });
+            // eslint-disable-next-line
             this.currentFrame = +frame;
             let style = this.currentModeId > 1 ? Object.assign({}, this.style[this.currentModeId]) : undefined;
             if (style && style.preset) delete style.preset;
