@@ -1,5 +1,11 @@
+import * as utils from './utils.js';
 import * as views from './views';
+import * as vue from './vue';
+import * as widgets from './widgets';
 
 export {
-    views
+    utils,
+    views,
+    vue,
+    widgets
 };
