@@ -1,9 +1,8 @@
-import $ from 'jquery';
-
-import Backbone from 'backbone';
-import {parseQueryString, splitRoute} from '@girder/core/misc';
-
 import ImageViewerWidget from './base';
+
+const $ = girder.$;
+const Backbone = girder.Backbone;
+const {parseQueryString, splitRoute} = girder.misc;
 
 // var baseSAUrl = 'https://unpkg.com/slideatlas-viewer@4.4.1/dist/';
 var baseSAUrl = 'https://cdn.jsdelivr.net/npm/slideatlas-viewer@4/dist/';
