@@ -1,6 +1,6 @@
-import Backbone from 'backbone';
-
 import ElementModel from '../models/ElementModel';
+
+const Backbone = girder.Backbone;
 
 export default Backbone.Collection.extend({
     initialize: function () {
