@@ -1,11 +1,19 @@
-from girder.plugin import GirderPlugin
+from pathlib import Path
+
+from girder.plugin import GirderPlugin, registerPluginStaticContent
 
 from . import assetstore
 
 
 class DICOMwebPlugin(GirderPlugin):
     DISPLAY_NAME = 'DICOMweb Plugin'
-    CLIENT_SOURCE_PATH = 'web_client'
 
     def load(self, info):
+        registerPluginStaticContent(
+            plugin='dicomweb',
+            css=[],
+            js=['/girder-plugin-dicomweb.umd.js'],
+            staticDir=Path(__file__).parent / 'web_client' / 'dist',
+            tree=info['serverRoot'],
+        )
         assetstore.load(info)
